post

+++ app/Http/Controllers/TasksController.php
@@ -15,11 +15,20 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
+        $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
 
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+            $data += $this->counts($user);
+            return view('users.show', $data);
+        }else {
+            return view('welcome');
+        }
     }
 
     /**
@@ -44,17 +53,18 @@ class TasksController extends Controller
      */
     public function store(Request $request)
     {
-        $this->validate($request, [
-            'status' => 'required|max:10',
-            'content' => 'required|max:10',
-        ]);
+        $id=\Auth::id();
         
-        $task = new Task;
-        $task ->status =$request->status;
-        $task->content = $request->content;
-        $task->save();
+        $this->validate($request, [
+            'content' => 'required|max:191',
+        ]);
 
-        return redirect('/');
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+        ]);
+
+        return redirect()->route("users.show", ["id"=>$id]);
     }
 
     /**
@@ -67,9 +77,15 @@ class TasksController extends Controller
     {
         $task = Task::find($id);
 
+        if (\Auth::id() === $task->user_id) {
         return view('tasks.show', [
             'task' => $task,
         ]);
+        }
+        
+        else {
+            return redirect("/");
+        }
     }
 
     /**
@@ -82,9 +98,13 @@ class TasksController extends Controller
     {
         $task = Task::find($id);
         
-        return view('tasks.edit',[
-            'task' =>$task,
-            ]);
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.edit',['task' =>$task]);
+        }
+        
+        else {
+                return redirect("/");
+        }
     }
 
     /**
@@ -102,11 +122,15 @@ class TasksController extends Controller
         ]);
         
         $task = Task::find($id);
+        
+        if (\Auth::id() === $task->user_id) {
+        
         $task->status = $request->status;
         $task->content = $request->content;
         $task->save();
+        }
         
-        return redirect('/');
+        return redirect('/users/'.\Auth::id());
     }
 
     /**
@@ -117,9 +141,14 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        $task = Task::find($id);
-        $task->delete();
+        $task = \App\Task::find($id);
 
-        return redirect('/');
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+            return redirect('/users/'.\Auth::id());
+        }
+        else {
+            return redirect('/');
+        }
     }
 }
